+++ craft_parts/plugins/jlink_plugin.py
@@ -94,10 +94,8 @@ class JLinkPlugin(Plugin):
     @override
     def get_build_commands(self) -> list[str]:
         """Return a list of commands to run during the build step."""
-        commands = []
-
-        # Set JAVA_HOME to be used in jlink commands
-        commands.append(
+        commands = [
+            # Set JAVA_HOME to be used in jlink commands
             """
                 if [ -z "${JAVA_HOME+x}" ]; then
                     JAVA_HOME=$(dirname $(dirname $(readlink -f $(which jlink))))
@@ -108,50 +106,42 @@ class JLinkPlugin(Plugin):
                 fi
                 JLINK=${JAVA_HOME}/bin/jlink
                 JDEPS=${JAVA_HOME}/bin/jdeps
-            """
-        )
-
-        # extract jlink version and use it to define the destination
-        # and multi-release jar version for the dependency enumeration
-        commands.append("JLINK_VERSION=$(${JLINK} --version)")
-        commands.append(
-            "DEST=usr/lib/jvm/java-${JLINK_VERSION%%.*}-openjdk-${CRAFT_ARCH_BUILD_FOR}"
-        )
-        commands.append("MULTI_RELEASE=${JLINK_VERSION%%.*}")
-
-        commands.append(self._get_find_jars_commands())
-
-        # create temp folder
-        commands.append("mkdir -p ${CRAFT_PART_BUILD}/tmp")
-        # extract jar files into temp folder - spring boot fat jar
-        # contains dependent jars inside
-        commands.append(
-            "(cd ${CRAFT_PART_BUILD}/tmp && for jar in ${PROCESS_JARS}; do jar xvf ${jar}; done;)"
-        )
-        # create classpath - add all dependent jars and all staged jars
-        commands.append("CPATH=.")
-        commands.append(
-            """
+            """,
+            # extract jlink version and use it to define the destination
+            # and multi-release jar version for the dependency enumeration
+            "JLINK_VERSION=$(${JLINK} --version)",
+            "DEST=usr/lib/jvm/java-${JLINK_VERSION%%.*}-openjdk-${CRAFT_ARCH_BUILD_FOR}",
+            "MULTI_RELEASE=${JLINK_VERSION%%.*}",
+            self._get_find_jars_commands(),
+            # Tell Jlink where to install.
+            "INSTALL_ROOT=${CRAFT_PART_INSTALL}/${DEST}",
+            # create temp folder
+            "mkdir -p ${CRAFT_PART_BUILD}/tmp",
+            # extract jar files into temp folder - spring boot fat jar
+            # contains dependent jars inside
+            "(cd ${CRAFT_PART_BUILD}/tmp && for jar in ${PROCESS_JARS}; do jar xvf ${jar}; done;)",
+            # create classpath - add all dependent jars and all staged jars
+            "CPATH=.",
+            """\
                 for file in $(find "${CRAFT_PART_BUILD}/tmp" -type f -name "*.jar"); do
                     CPATH="$CPATH:${file}"
                 done
                 for file in $(find "${CRAFT_STAGE}" -type f -name "*.jar"); do
                     CPATH="$CPATH:${file}"
                 done
-            """
-        )
-        commands.append(
-            """if [ "x${PROCESS_JARS}" != "x" ]; then
-                deps=$(${JDEPS} --print-module-deps -q --recursive \
-                    --ignore-missing-deps \
-                    --multi-release ${MULTI_RELEASE} \
-                    --class-path=${CPATH} \
-                    ${PROCESS_JARS})
+            """,
+            """\
+                if [ "x${PROCESS_JARS}" != "x" ]; then
+                    deps=$(${JDEPS} --print-module-deps -q --recursive \
+                      --ignore-missing-deps \
+                      --multi-release ${MULTI_RELEASE} \
+                      --class-path=${CPATH} \
+                      ${PROCESS_JARS})
                 else
                     deps=java.base
                 fi
-            """
-        )
+            """,
+        ]
         extra_modules = self._get_extra_module_list()
         if len(extra_modules) > 0:
             commands.append(f"deps=${{deps}},{extra_modules}")